refactored chatWidget + resized component
