+++ vite-project/src/Modules/Chat/ChatWidget/Components/ChatWidget.css
@@ -23,7 +23,7 @@
 
 .chat-window {
   background-color: #fff;
-  width: 350px;
+  width: 400px;
   height: 400px;
   box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
   display: flex;
@@ -43,10 +43,9 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  cursor: pointer;
 }
 
-.chat-header button {
+.chat-header-icon {
   background: none;
   border: none;
   color: #fff;
